Final Project Changes
