+++ USF_Library/Staff.cs
@@ -1,0 +1,17 @@
+using System;
+namespace USF_Library 
+{
+    public class Staff : Person
+    {
+        public string Position { get; set; }
+        public string Department { get; set; }
+
+        public Staff(string name, string email, string id, string position, string department)
+            : base(name, email, id)
+        {
+            Position = position;
+            Department = department;
+        }
+    }
+
+}
