+++ USF_Library/bin/Debug/net8.0/USF_Library.deps.json
@@ -1,0 +1,23 @@
+{
+  "runtimeTarget": {
+    "name": ".NETCoreApp,Version=v8.0",
+    "signature": ""
+  },
+  "compilationOptions": {},
+  "targets": {
+    ".NETCoreApp,Version=v8.0": {
+      "USF_Library/1.0.0": {
+        "runtime": {
+          "USF_Library.dll": {}
+        }
+      }
+    }
+  },
+  "libraries": {
+    "USF_Library/1.0.0": {
+      "type": "project",
+      "serviceable": false,
+      "sha512": ""
+    }
+  }
+}
